requirements.txt
atproto>=0.0.55
weasyprint>=62.0
python-dateutil>=2.8.2
Jinja2>=3.1.2
httpx>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
